use standard FHIR event schema
