+++ src/operations/common/logging.js
@@ -1,8 +1,11 @@
 const env = require('var');
+const moment = require('moment-timezone');
 /**
  * @type {import('winston').logger}
  */
 const logger = require('@asymmetrik/node-fhir-server-core').loggers.get();
+
+const fhirLogger = require('../../utils/fhirLogger').FhirLogger.getLogger();
 
 /**
  * Always logs regardless of env.IS_PRODUCTION
@@ -40,4 +43,77 @@ module.exports.logError = (user, msg) => {
  */
 module.exports.logWarn = (user, msg) => {
     logger.warn(user + ': ' + msg);
+};
+
+/**
+ * Logs a FHIR operation
+ * @param {import('../../utils/requestInfo').RequestInfo} requestInfo
+ * @param {Object} args
+ * @param {string | null} scope
+ * @param {string} resourceType
+ * @param {number|null} startTime
+ * @param {number|null} stopTime
+ * @param {string} message
+ */
+module.exports.logOperation = (requestInfo, args,
+                               scope, resourceType,
+                               startTime, stopTime,
+                               message) => {
+    /**
+     * @type {{valueString: string|undefined, valuePositiveInt: number|undefined, type: string}[]}
+     */
+    let detail = Object.entries(args).map(([k, v]) => {
+            return {
+                type: k,
+                valueString: String(v)
+            };
+        }
+    );
+    if (startTime && stopTime) {
+        /**
+         * @type {number}
+         */
+        const elapsedSeconds = stopTime - startTime;
+        detail.push({
+            type: 'duration',
+            valuePositiveInt: elapsedSeconds
+        });
+    }
+    fhirLogger.info(
+        {
+            id: requestInfo.requestId,
+            type: {
+                code: 'fhirServer'
+            },
+            action: 'searchStreaming',
+            period: {
+                start: new Date(startTime).toISOString(),
+                end: new Date(stopTime).toISOString(),
+            },
+            recorded: new Date(moment.utc().format('YYYY-MM-DDTHH:mm:ssZ')),
+            outcome: 0, // https://hl7.org/fhir/valueset-audit-event-outcome.html
+            outcomeDesc: 'Success',
+            agent: [
+                {
+                    altId: (typeof requestInfo.user === 'string') ? requestInfo.user : requestInfo.user.id,
+                    network: {
+                        address: requestInfo.remoteIpAddress
+                    },
+                    policy: [
+                        scope
+                    ]
+                }
+            ],
+            source: {
+                site: requestInfo.originalUrl
+            },
+            entity: [
+                {
+                    name: resourceType,
+                    detail: detail
+                }
+            ],
+            message: message
+        }
+    );
 };
